+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Enhanced `Context` class with the ability to update thread, messages, and run details.
 - Create DatabaseManagement class to handle database interact function
 - Enable thread worker count customization
+- Enable worker's task timeout customization
 
 ## [0.0.3] - 14/01/2025